Vrty little actual progress. needed to backtrack, created two seperate methods for save and create.
now create does nothing (nearly) it
creates a document with uid but no fields.

delete button still not working and no leads as to what could b
e the cause.

Hopefully tomorrow can be a more productive day for this project and
B E Y O N D

+++ src/app/create-thread/create-thread.component.ts
@@ -26,14 +26,22 @@ export class CreateThreadComponent implements OnInit {
 
   }
 
-    save(thread) {
-        if(this.id) this.threadService.update(this.id, thread);
-        else this.threadService.create(thread);
-        this.router.navigate(['/forum/view-threads'])
+    save(thread, uid=null) {
+        if (uid) {
+            this.threadService.save(thread, uid);
+        } else {
+            //could have this method return the uid and then cache it somehow
+            this.threadService.add(thread);
+        }
+        
+        this.router.navigate(['/forum/view-threads']);
     }
 
     delete(id) {
-        this.threadService.delete(id);
+        if (confirm("Are you sure you want to delete this thread? (cannot be undone)")){
+            this.threadService.delete(id);
+            this.router.navigate(['/forum/view-threads']);}
+        else return 0;
     }
